Tighten account refill review edges
CodeRabbit review pointed out that monetary refill DTOs should avoid primitive doubles, locking queries should be explicit, and purchase/refill lock assumptions should be easier to verify. This commit keeps the approved lock order while making those contracts sharper in code, tests, and product docs.

Constraint: Purchase must still lock the daily refill state before reward rows to avoid account/reward/refill deadlocks.

Rejected: Move refill credit after reward row locks | that reintroduces the lock-order cycle the architect review caught.

Confidence: high

Scope-risk: narrow

Directive: AccountRefillCreditProcessor uses MANDATORY transaction propagation so purchase rollback also rolls back the credited refill count.

Tested: backend ./gradlew check --console=plain

Tested: frontend npm run lint; npm test; npm run build

+++ backend/src/main/java/coin/coinzzickmock/feature/account/application/result/AccountRefillResult.java
@@ -1,8 +1,10 @@
 package coin.coinzzickmock.feature.account.application.result;
 
+import java.math.BigDecimal;
+
 public record AccountRefillResult(
-        double walletBalance,
-        double availableMargin,
+        BigDecimal walletBalance,
+        BigDecimal availableMargin,
         int remainingCount
 ) {
 }
