+++ backend/src/main/java/coin/coinzzickmock/feature/account/application/result/AccountRefillStatusResult.java
@@ -1,13 +1,14 @@
 package coin.coinzzickmock.feature.account.application.result;
 
+import java.math.BigDecimal;
 import java.time.Instant;
 
 public record AccountRefillStatusResult(
         int remainingCount,
         boolean refillable,
         String disabledReason,
-        double targetWalletBalance,
-        double targetAvailableMargin,
+        BigDecimal targetWalletBalance,
+        BigDecimal targetAvailableMargin,
         Instant nextResetAt
 ) {
 }
