+++ backend/src/main/java/coin/coinzzickmock/feature/account/web/AccountRefillStatusResponse.java
@@ -1,13 +1,14 @@
 package coin.coinzzickmock.feature.account.web;
 
+import java.math.BigDecimal;
 import java.time.Instant;
 
 public record AccountRefillStatusResponse(
         int remainingCount,
         boolean refillable,
         String disabledReason,
-        double targetWalletBalance,
-        double targetAvailableMargin,
+        BigDecimal targetWalletBalance,
+        BigDecimal targetAvailableMargin,
         Instant nextResetAt
 ) {
 }
